Scalar truncation size: 512 -> 256
This creates a bit of symmetry with the scalar sample size (64 bytes)
and ref key sample count (4).

I think this is still reasonable, but I suspect that rather than trying
to tune this size to be exactly right it'd be better to support
a 'slice'-ish syntax for people to be able to pick chunks out of strings
or arrays.