Prototype: Perl-ish syntax for accessing vars in scope
This is potentially a really terrible idea, but it was fun to work on
the parser. Basically: JSON::Pointer syntax is probably a little too
simple -- Sawyer had good feedback on not just guessing whether
something is a hash or array access. So he wrote a nice, compact lexer
and parser to support quoted strings and spit out tokens with types.

This commit takes that general idea, and goes a little nuts: the parser
now produces an AST describing a Perl-like syntax for dereferencing and
accessing Perl data structures. It supports simple cases, like
'$my_cool_hashref', but also fancy ones, like
'$my_cool_hashref->{$some_arrayref->[-1]->{nested_hashkey}->{$some_scalar}}';

It does not use Perl 'eval'; indeed, it violates normal Perl syntax
rules in several ways. For example, indexing into a data structure
_always_ uses '->', even if it isn't a ref, e.g. '%my_cool_hash->{foo}'.
As a result, this might be uncanny valley territory.

TODO: tests, the interpreter, reviewing whether this is remotely
a reasonable thing to do.