interpreter: naive recursive interpreter
This commit introduces a basic recursive interpreter for the AST spat
out by the parser. With this, the basic functionality of Devel::Optic is
back in place, now with a much higher power language for data access.

Ideally I'd like to change this and the parser to be non-recursive to
reduce the number of sub calls -- the interpeter code, at least, might
be running inside very hot parts of a system -- but I couldn't quite
picture how to do this with explicit stacks, and I wanted to get
something functioning to start with. Make it work, make it right, make
it fast, yada yada yada.

TODO: adapt tests, adapt main functions to use this, more human readable
error messages, performance measurements and optimizations.

The error messages are going to be quite a challenge -- the simple (and
non-nested) JSON::Pointer syntax made it easy to tell users exactly
where we were when things went pear shaped, so we could give very useful
feedback to guide iteration. That's a bit trickier with recursive tree
traversal.