Remove Devel::Size.
Some profiling indicated that it was by far the most expensive part of
Devel::Optic. With it enabled, throughput on a "hello world" Mojolicious
app dropped to about 50% of normal. Without it, overhead was small
enough to be nearly unmeasureable.

This has the downside that Devel::Optic now _always_ summarizes.

A better conceptual approach would be to write a custom serializer:
start traversing the data structure and adding to the serialized
representation. When that representation becomes too big, stop and send
it. That would get maximum information for minimum size. However, the
current strategy (always do a basic one-level summary) is much simpler
to implement, so we'll start with this and await some feedback.